Launch user goroutin in controller
