+++ pkg/main.go
@@ -2,8 +2,6 @@ package main
 
 import (
 	"climb/pkg/controller"
-	"climb/pkg/types"
-	"climb/pkg/utils"
 	"context"
 	"log"
 	"os"
@@ -65,72 +63,11 @@ func main() {
 	u.Timeout = 60
 
 	updates, err := bot.GetUpdatesChan(u)
-
-	// Prepare a map of user -> chan update
-	userForwarder := make(map[int]chan tgbotapi.Update)
+	if err != nil {
+		log.Println(err.Error())
+	}
 
 	for update := range updates {
-
-		userId := utils.GetUser(&update).ID
-
-		channel, prs := userForwarder[userId]
-		if !prs {
-			newChannel := make(chan tgbotapi.Update)
-			userForwarder[userId] = newChannel
-			channel = newChannel
-
-			go handleUser(
-				controller,
-				channel,
-			)
-		}
-
-		channel <- update
-	}
-}
-
-func handleUser(
-	ctrl controller.Controller,
-	updates <-chan tgbotapi.Update,
-) {
-	var forwarder *types.Comm
-	commandTermination := make(chan interface{})
-
-	for {
-		select {
-		case <-commandTermination:
-			// commands wants to end
-			if forwarder != nil {
-				forwarder.StopCommand <- struct{}{}
-				forwarder = nil
-			}
-			break
-
-		case update := <-updates:
-			// we received an update message from our user
-			utils.LogReception(update)
-
-			if update.Message != nil && update.Message.IsCommand() {
-
-				// Clean up previous commands
-				if forwarder != nil {
-					forwarder.StopCommand <- struct{}{}
-					forwarder = nil
-				}
-
-				// Get new command started
-				for _, cmd := range ctrl.AvailableCommands() {
-					if update.Message.Command() == cmd.Command {
-						comm := cmd.Instantiation(commandTermination)
-						forwarder = &comm
-						forwarder.Updates <- update
-						break
-					}
-				}
-			} else if forwarder != nil {
-				forwarder.Updates <- update
-			}
-			break
-		}
+		controller.GetAssociatedChan(update) <- update
 	}
 }
